Project Initialization & Package Inports
This is a blank Universal 2D project made with Unity 6.3.2 . The project is using the Universal Render Pipeline (URP) and I have already imported numerous useful packages, notably; Odin Inspector, Cinemachine, Netcode For GameObjects. For the full list of installed packages, please refer to the Package Manager window in the project's Unity Editor.